Refactor platform routes
