+++ src/web/mod.rs
@@ -11,7 +11,7 @@ pub use health_routes::router as health_routes;
 pub use login_routes::router as login_routes;
 pub use oauth_routes::router as oauth_routes;
 pub use openai::router as openai_routes;
-pub use platform::org_routes as platform_org_routes;
+pub use platform::router as platform_routes;
 pub use protected_routes::router as protected_routes;
 
 pub use platform::login_routes as platform_login_routes;
