+++ src/web/oauth_routes.rs
@@ -2,8 +2,8 @@ use crate::decrypt_with_key;
 use crate::models::email_verification::NewEmailVerification;
 use crate::models::oauth::NewUserOAuthConnection;
 use crate::oauth::OAuthState;
-use crate::platform_org_routes::PROJECT_GITHUB_OAUTH_SECRET;
-use crate::platform_org_routes::PROJECT_GOOGLE_OAUTH_SECRET;
+use crate::web::platform::PROJECT_GITHUB_OAUTH_SECRET;
+use crate::web::platform::PROJECT_GOOGLE_OAUTH_SECRET;
 use crate::web::encryption_middleware::{decrypt_request, encrypt_response, EncryptedResponse};
 use crate::web::login_routes::handle_new_user_registration;
 use crate::GithubProvider;
