+++ src/web/platform/common.rs
@@ -1,0 +1,197 @@
+use crate::{
+    models::{org_memberships::OrgRole, project_settings::OAuthProviderSettings},
+    web::platform::validation::{
+        validate_alphanumeric_only, validate_alphanumeric_with_symbols, validate_secret_size,
+    },
+};
+use chrono::{DateTime, Utc};
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+use validator::Validate;
+
+pub const PROJECT_RESEND_API_KEY: &str = "RESEND_API_KEY";
+pub const PROJECT_GOOGLE_OAUTH_SECRET: &str = "GOOGLE_OAUTH_SECRET";
+pub const PROJECT_GITHUB_OAUTH_SECRET: &str = "GITHUB_OAUTH_SECRET";
+
+// Request Types
+#[derive(Deserialize, Clone, Validate)]
+pub struct CreateOrgRequest {
+    #[validate(length(min = 1, max = 50))]
+    #[validate(custom(function = "validate_alphanumeric_with_symbols"))]
+    pub name: String,
+}
+
+#[derive(Deserialize, Clone, Validate)]
+pub struct CreateProjectRequest {
+    #[validate(length(min = 1, max = 50))]
+    #[validate(custom(function = "validate_alphanumeric_with_symbols"))]
+    pub name: String,
+    #[validate(length(max = 255))]
+    pub description: Option<String>,
+}
+
+#[derive(Deserialize, Clone, Validate)]
+pub struct UpdateProjectRequest {
+    #[validate(length(min = 1, max = 50))]
+    #[validate(custom(function = "validate_alphanumeric_with_symbols"))]
+    pub name: Option<String>,
+    #[validate(length(max = 255))]
+    pub description: Option<String>,
+    #[validate(custom(function = "validate_project_status"))]
+    pub status: Option<String>,
+}
+
+#[derive(Deserialize, Clone, Validate)]
+pub struct CreateInviteRequest {
+    #[validate(email(message = "Invalid email format"))]
+    #[validate(length(max = 255, message = "Email must not exceed 255 characters"))]
+    pub email: String,
+    #[serde(default = "default_invite_role")]
+    pub role: OrgRole,
+}
+
+#[derive(Deserialize, Clone, Validate)]
+pub struct UpdateMembershipRequest {
+    pub role: OrgRole,
+}
+
+#[derive(Deserialize, Clone, Validate)]
+pub struct CreateSecretRequest {
+    #[validate(length(min = 1, max = 50))]
+    #[validate(custom(function = "validate_alphanumeric_only"))]
+    pub key_name: String,
+    #[validate(custom(function = "validate_secret_size"))]
+    pub secret: String, // Base64 encoded secret value
+}
+
+#[derive(Deserialize, Clone, Validate)]
+pub struct UpdateEmailSettingsRequest {
+    #[validate(length(min = 1, max = 255))]
+    #[validate(custom(function = "validate_email_provider"))]
+    pub provider: String,
+    #[validate(email)]
+    pub send_from: String,
+    pub email_verification_url: String,
+}
+
+#[derive(Deserialize, Clone, Validate)]
+pub struct UpdateOAuthSettingsRequest {
+    pub google_oauth_enabled: bool,
+    pub github_oauth_enabled: bool,
+    #[validate(custom(function = "validate_oauth_provider_settings"))]
+    pub google_oauth_settings: Option<OAuthProviderSettings>,
+    #[validate(custom(function = "validate_oauth_provider_settings"))]
+    pub github_oauth_settings: Option<OAuthProviderSettings>,
+}
+
+// Response Types
+#[derive(Serialize)]
+pub struct OrgResponse {
+    pub id: Uuid,
+    pub name: String,
+}
+
+#[derive(Serialize)]
+pub struct ProjectResponse {
+    pub id: Uuid,
+    pub client_id: Uuid,
+    pub name: String,
+    pub description: Option<String>,
+    pub status: String,
+    pub created_at: DateTime<Utc>,
+}
+
+#[derive(Serialize)]
+pub struct MembershipResponse {
+    pub user_id: Uuid,
+    pub role: String,
+    pub name: Option<String>,
+}
+
+#[derive(Serialize)]
+pub struct InviteResponse {
+    pub code: Uuid,
+    pub email: String,
+    pub role: String,
+    pub used: bool,
+    pub expires_at: DateTime<Utc>,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+}
+
+#[derive(Serialize)]
+pub struct DetailedInviteResponse {
+    pub code: Uuid,
+    pub email: String,
+    pub role: String,
+    pub used: bool,
+    pub expires_at: DateTime<Utc>,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+    pub organization_name: String,
+}
+
+#[derive(Serialize)]
+pub struct SecretResponse {
+    pub key_name: String,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+}
+
+#[derive(Debug, Serialize)]
+pub struct PlatformUserResponse {
+    pub id: Uuid,
+    pub email: String,
+    pub name: Option<String>,
+    pub email_verified: bool,
+    pub created_at: DateTime<Utc>,
+    pub updated_at: DateTime<Utc>,
+}
+
+#[derive(Serialize)]
+pub struct MeResponse {
+    pub user: PlatformUserResponse,
+    pub organizations: Vec<OrgResponse>,
+}
+
+// Validation Functions
+pub fn validate_project_status(status: &str) -> Result<(), validator::ValidationError> {
+    match status {
+        "active" | "inactive" | "suspended" => Ok(()),
+        _ => Err(validator::ValidationError::new("project_status")),
+    }
+}
+
+pub fn default_invite_role() -> OrgRole {
+    OrgRole::Admin
+}
+
+pub fn validate_email_provider(provider: &str) -> Result<(), validator::ValidationError> {
+    if provider != "resend" {
+        let mut error = validator::ValidationError::new("invalid_email_provider");
+        error.message = Some("Only 'resend' is supported as an email provider".into());
+        return Err(error);
+    }
+    Ok(())
+}
+
+pub fn validate_oauth_provider_settings(
+    settings: &&OAuthProviderSettings,
+) -> Result<(), validator::ValidationError> {
+    // Validate client_id
+    if settings.client_id.is_empty() || settings.client_id.len() > 255 {
+        return Err(validator::ValidationError::new("oauth_client_id"));
+    }
+    // Validate redirect_url
+    if settings.redirect_url.is_empty() || settings.redirect_url.len() > 255 {
+        return Err(validator::ValidationError::new("oauth_redirect_url"));
+    }
+    // Basic URL validation
+    if url::Url::parse(&settings.redirect_url).is_err() {
+        return Err(validator::ValidationError::new(
+            "oauth_redirect_url_invalid",
+        ));
+    }
+    Ok(())
+}
+
